+++ app/models/sheet.rb
@@ -2,6 +2,24 @@
 
 # Defines a collection of responses to a design for a subject.
 class Sheet < ApplicationRecord
+  ORDERS = {
+    'site' => 'sites.name',
+    'site desc' => 'sites.name desc',
+    'design' => 'designs.name',
+    'design desc' => 'designs.name desc',
+    'created_by' => 'users.last_name, users.first_name',
+    'created_by desc' => 'users.last_name desc nulls last, users.first_name desc nulls last',
+    'subject' => 'subjects.subject_code',
+    'subject desc' => 'subjects.subject_code desc',
+    'percent' => 'sheets.percent',
+    'percent desc' => 'sheets.percent desc nulls last',
+    'created' => 'sheets.created_at',
+    'created desc' => 'sheets.created_at desc',
+    'edited' => 'sheets.last_edited_at',
+    'edited desc' => 'sheets.last_edited_at desc nulls last'
+  }
+  DEFAULT_ORDER = 'sheets.last_edited_at desc nulls last'
+
   # Concerns
   include Deletable, Latexable, Siteable, Evaluatable, AutoLockable, Forkable, Coverageable
 
@@ -11,9 +29,6 @@ class Sheet < ApplicationRecord
   scope :search, -> (arg) { where('sheets.subject_id in (select subjects.id from subjects where subjects.deleted = ? and LOWER(subjects.subject_code) LIKE ?) or design_id in (select designs.id from designs where designs.deleted = ? and LOWER(designs.name) LIKE ?)', false, arg.to_s.downcase.gsub(/^| |$/, '%'), false, arg.to_s.downcase.gsub(/^| |$/, '%')).references(:designs) }
   scope :sheet_before, -> (*args) { where('sheets.created_at < ?', (args.first + 1.day).at_midnight) }
   scope :sheet_after, -> (*args) { where('sheets.created_at >= ?', args.first.at_midnight) }
-
-  scope :with_variable_response, -> (*args) { where('sheets.id IN (select sheet_variables.sheet_id from sheet_variables where sheet_variables.variable_id = ? and sheet_variables.value = ?)', args.first, args[1]) }
-  scope :with_checkbox_variable_response, -> (*args) { where('sheets.id IN (select responses.sheet_id from responses where responses.variable_id = ? and responses.value = ? )', args.first, args[1]) }
 
   # These don't include blank codes
   scope :with_variable_response_after, -> (*args) { where("sheets.id IN (select sheet_variables.sheet_id from sheet_variables where sheet_variables.variable_id = ? and sheet_variables.value >= ? and sheet_variables.value != '')", args.first, args[1]) }
@@ -100,34 +115,34 @@ class Sheet < ApplicationRecord
 
   # stratum can be nil (grouping on site) or a variable (grouping on the variable responses)
   # TODO: This can be cleaned up using the new Search module along with operators.
-  def self.with_stratum(current_user, variable, stratum_value, operator, stratum_start_date = nil, stratum_end_date = nil)
+  def self.with_stratum(current_user, variable, value, operator, stratum_start_date = nil, stratum_end_date = nil)
     if variable.variable_type == 'design'
-      where(design_id: stratum_value)
+      where(design_id: value)
     elsif variable.variable_type == 'site'
-      with_site(stratum_value)
+      with_site(value)
     elsif operator == 'any' && !%w(sheet_date).include?(variable.variable_type)
       filter_variable(variable, current_user, 'any')
     elsif %w(sheet_date date).include?(variable.variable_type) && !%w(blank missing).include?(operator)
       sheet_after_variable(variable, stratum_start_date).sheet_before_variable(variable, stratum_end_date)
-    elsif stratum_value.present? # Ex: variable: variables(:gender), stratum_value: 'f'
+    elsif value.present? # Ex: variable: variables(:gender), value: 'f'
       if variable.variable_type == 'file'
         # TODO: This may be able to target a specific file.
         filter_variable(variable, current_user, 'any')
       elsif variable.variable_type == 'checkbox'
-        with_checkbox_variable_response(variable, stratum_value)
+        filter_variable(variable, current_user, '=', value: value)
       else
-        with_variable_response(variable, stratum_value)
+        filter_variable(variable, current_user, '=', value: value)
       end
     elsif operator == 'blank'
       filter_variable(variable, current_user, operator)
-    else # Ex: variable: variables(:gender), stratum_value: nil
+    else # Ex: variable: variables(:gender), value: nil
       filter_variable(variable, current_user, 'missing')
     end
   end
 
   # TODO: Temporary rewrite to use Search instead of sheet scopes
-  def self.filter_variable(variable, current_user, operator)
-    token = Token.new(key: variable.name, operator: operator, variable: variable)
+  def self.filter_variable(variable, current_user, operator, value: nil)
+    token = Token.new(key: variable.name, operator: operator, variable: variable, value: value)
     Search.run_sheets(
       variable.project,
       current_user,